feat: share plugin loading between run and TUI commands
Extract initializePlugins() helper in plugin/loader.ts that loads,
merges, sets registry, and logs. Use it from both run-command and
tui-command so TUI sessions also have access to plugins.
