+++ packages/runtime/src/cli/tui-command.ts
@@ -8,6 +8,7 @@ import { initDb } from "$/db/index.js";
 import { flushAllSessions, loadSessions } from "$/db/sessions.js";
 import colors from "$/output/colors.js";
 import { error, config, setLogFile } from "$/output/log.js";
+import { initializePlugins } from "$/plugin/loader.js";
 import { root } from "$/util/paths.js";
 import { onShutdown, registerSigint } from "$/util/shutdown.js";
 import { input, select } from "@inquirer/prompts";
@@ -19,6 +20,7 @@ async function run(_noFlags: {}, agentSlug: string): Promise<void> {
   setLogFile(path.join(root(), "logs", "cireilclaw.log"));
 
   await runMigrations();
+  await initializePlugins();
 
   registerSigint();
   onShutdown(() => {
